requirements.txt
aio-pika==9.4.0
fastapi==0.109.2
pydantic==2.6.1
pydantic-settings==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
uvicorn[standard]==0.27.0.post1
python-dotenv~=1.0.1